Remove useless code_model indirection
